mises à 0-100 Limites

+++ Tamagotchi/src/vue/PanelProgressBar.java
@@ -76,11 +76,11 @@ public class PanelProgressBar extends JPanel {
 
     public void actualiserBarresSelonConstantes(Tamagotchi tama) {
         // Met les barres aux valeurs des attributs du tamagotchi
-        setValueBarreVie((int) tama.getVie());
-        setValueBarreNourriture((int) tama.getFaim());
-        setValueBarreSommeil((int) tama.getSommeil());
-        setValueBarreHygiene((int) tama.getHygiene());
-        setValueBarreLoisir((int) tama.getLoisir());
+        setValueBarreVie(tama.getVie());
+        setValueBarreNourriture(tama.getFaim());
+        setValueBarreSommeil(tama.getSommeil());
+        setValueBarreHygiene(tama.getHygiene());
+        setValueBarreLoisir(tama.getLoisir());
         this.revalidate();
     }
